+++ lib/hatchet/tasks.rb
@@ -1,7 +1,11 @@
 namespace :hatchet do
   task :setup_ci do
     script = File.expand_path(File.join(__dir__, "../../etc/ci_setup.rb"))
-    out = `#{script}`
+    puts "Running script `#{script}`"
+    out = `#{script} 2>&1`
+    puts "Done"
+    puts "Output:\n"
+    puts out
     raise "Command #{script.inspect} failed\n#{out}" unless $?.success?
   end